new commit week-8 task

+++ src/main/java/com/employee/employee_management_system/controller/EmployeeController.java
@@ -1,36 +1,46 @@
 package com.employee.employee_management_system.controller;
 
 import com.employee.employee_management_system.model.Employee;
-import com.employee.employee_management_system.model.SalaryRecord;
 import com.employee.employee_management_system.services.EmployeeService;
+import com.employee.employee_management_system.services.SalaryService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.*;
 
+import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpSession;
-import java.util.List;
 
 
 @Controller
 public class EmployeeController {
 
-    private EmployeeService employeeService;
+    private final EmployeeService employeeService;
+    private final SalaryService salaryService;
 
     @Autowired
 
-    public EmployeeController(EmployeeService employeeService) {
+    public EmployeeController(EmployeeService employeeService, SalaryService salaryService) {
         this.employeeService = employeeService;
+        this.salaryService = salaryService;
     }
 
     @GetMapping("/employee-list")
-    public String viewHomePage(Model model){
+    public String viewHomePage(Model model, HttpServletRequest request){
+        HttpSession session = request.getSession(false);
+        if (session == null || session.getAttribute("admin") == null) {
+            return "redirect:/";
+        }
+
         model.addAttribute("listEmployees", employeeService.getAllEmployees());
         return "frontPage";
     }
 
     @GetMapping(path = "/new-employee")
-    public String displayEmployeeForm(Model model, HttpSession session){
+    public String displayEmployeeForm(Model model, HttpSession session, HttpServletRequest request){
+        Object userObj = session.getAttribute("admin");
+        if (userObj == null) return "redirect:/";
+
         model.addAttribute("invalid", session.getAttribute("invalid"));
         session.removeAttribute("invalid");
         model.addAttribute("employee", new Employee());
@@ -40,31 +50,75 @@ public class EmployeeController {
     @PostMapping("/new-employee")
     public String addEmployee(@ModelAttribute("employee") Employee employee, HttpSession session,Model model){
 
-//        Employee employee1 = (Employee) model.getAttribute("employee");
+        Object userObj = session.getAttribute("admin");
+        if (userObj == null) return "redirect:/";
+
         employeeService.createEmployee(employee);
         session.setAttribute("newemployee", employee);
         return "redirect:/employee-list";
     }
 
-
     @GetMapping("/update-employee/{id}")
-    public String showEditEmployee(@PathVariable (value = "id") Long id, Model model) {
-        model.addAttribute("employee", employeeService.findEmployee(id));
+    public String showFormForUpdate(@PathVariable( value = "id") long id, Model model, HttpSession session){
+        Object userObj = session.getAttribute("admin");
+        if (userObj == null) return "redirect:/";
+
+//        //set employee as a model attribute to pre-populate the form
+        Employee employee = employeeService.findEmployee(id);
+        model.addAttribute("employee", employee);
         return "updateEmployee";
     }
 
-    @RequestMapping("/delete-employee/{id}")
-    public String deleteEmployee(@PathVariable(value = "id") long id, Model model) {
+    @PostMapping("/update-employee/{id}")
+    public String showEditEmployee(@ModelAttribute("employee") Employee employee, @PathVariable (value = "id") Long id, Model model, HttpSession session) {
 
+        Object userObj = session.getAttribute("admin");
+        if (userObj == null) return "redirect:/";
+
+        //save employee to database
+        employeeService.updateEmployee(employee, id);
+        return "redirect:/employee-list";
+    }
+
+    @RequestMapping("/delete-employee/{id}")
+    public String deleteEmployee(@PathVariable(value = "id") long id, Model model, HttpSession session) {
+        Object userObj = session.getAttribute("admin");
+        if (userObj == null) return "redirect:/";
         employeeService.deleteEmployeeById(id);
         return "redirect:/";
     }
 
-//    @GetMapping("/{id}")
-//    public String viewEmployee(@PathVariable(value="id") Long id, Model model){
+    @GetMapping("/employee-list/{id}")
+    public String viewEmployee(@PathVariable(value="id") Long id, Model model, HttpSession session){
+        Object userObj = session.getAttribute("admin");
+        if (userObj == null) return "redirect:/";
+        Employee employee = employeeService.findEmployee(id);
+
+        model.addAttribute("employee", employee);
+        model.addAttribute("salaries", salaryService.getEmployeeSalaries(id));
+        return "employeeInfo";
+    }
+
+//    @GetMapping("/employee")
+//    public String employeePage(Model model, HttpServletRequest request){
+//        HttpSession session = request.getSession(false);
+//        if (session == null || session.getAttribute("employee") == null) {
+//            return "redirect:/";
+//        }
+//
+//        model.addAttribute("employee", employeeService.getAllEmployees());
+//        return "plainEmployeeInfo";
+//    }
+
+//    @GetMapping("/employee/{id}")
+//    public String employee(@PathVariable(value="id") Long id, Model model, HttpSession session){
+//        Object userObj = session.getAttribute("employee");
+//        if (userObj == null) return "redirect:/";
+//
 //        Employee employee = employeeService.findEmployee(id);
-//        List<SalaryRecord> salaryHistory = employeeService.getEmployeeSalaryHistory(employee);
+//
 //        model.addAttribute("employee", employee);
-//        return "employeeInfo";
+//        model.addAttribute("salaries", salaryService.getEmployeeSalaries(id));
+//        return "plainEmployeeInfo";
 //    }
 }
